New options in the configuration file:
[settings]
mode = live

The mode can be set to 'test' to use the build in test_mode features, or to run live for the production work.

Also the post_install script is touched and is pretty empty at the moment.

A new Arista specific repository has been created that is now REQUIRED to be used at least once before installing the RPM/Extension.

To be found here: https://github.com/okoeroo/arista-EOS-rc
Install an "eos.rc" file and "eos.rc.d/" on the flash disk to aid 3rd party software developers on the EOS platform.